+++ src/extension.ts
@@ -1,17 +1,22 @@
 import * as vscode from 'vscode'
+import ProfileService from './core/services/profileService'
 
-export async function activate(context: vscode.ExtensionContext) {
+export async function activate(ctx: vscode.ExtensionContext) {
 	// Register each command individually
-	context.subscriptions.push(
-		vscode.commands.registerCommand(`xx.menu`, async () => (await import('./commands/menuTest')).cmd.foo(context)),
-		vscode.commands.registerCommand(`xx.createProfile`, async () => (await import('./commands/createProfile')).cmd.foo(context)),
-		vscode.commands.registerCommand(`xx.deleteProfile`, async () => (await import('./commands/deleteProfile')).cmd.foo(context)),
-		vscode.commands.registerCommand(`xx.duplicateProfile`, async () => (await import('./commands/duplicateProfile')).cmd.foo(context)),
-		vscode.commands.registerCommand(`xx.editProfile`, async () => (await import('./commands/editProfile')).cmd.foo(context)),
-		vscode.commands.registerCommand(`xx.enableProfile`, async () => (await import('./commands/enableProfile')).cmd.foo(context)),
-		vscode.commands.registerCommand(`xx.mergeProfile`, async () => (await import('./commands/mergeProfile')).cmd.foo(context)),
-		vscode.commands.registerCommand(`xx.renameProfiles`, async () => (await import('./commands/renameProfiles')).cmd.foo(context)),
+	ctx.subscriptions.push(
+		vscode.commands.registerCommand(`xx.menu`, async () => (await import('./commands/menuTest')).cmd.foo(ctx)),
+		vscode.commands.registerCommand(`xx.createProfile`, async () => (await import('./commands/createProfile')).cmd.foo(ctx)),
+		vscode.commands.registerCommand(`xx.deleteProfile`, async () => (await import('./commands/deleteProfile')).cmd.foo(ctx)),
+		vscode.commands.registerCommand(`xx.duplicateProfile`, async () => (await import('./commands/duplicateProfile')).cmd.foo(ctx)),
+		vscode.commands.registerCommand(`xx.editProfile`, async () => (await import('./commands/editProfile')).cmd.foo(ctx)),
+		vscode.commands.registerCommand(`xx.enableProfile`, async () => (await import('./commands/enableProfile')).cmd.foo(ctx)),
+		vscode.commands.registerCommand(`xx.mergeProfile`, async () => (await import('./commands/mergeProfile')).cmd.foo(ctx)),
+		vscode.commands.registerCommand(`xx.renameProfiles`, async () => (await import('./commands/renameProfiles')).cmd.foo(ctx)),
 	)
+
+	// Create an empty global profile if one doesn't exist
+	if (!ProfileService.profiles(ctx).find(p => p.name == 'Global'))
+		await ProfileService.createProfile('Global', [], [], ctx)
 }
 
 export function deactivate() { }